Rendr class refactor: basic shader + projection matrix

+++ header/Generator.h
@@ -8,14 +8,14 @@ class Generator {
 public:
     Generator();
     virtual ~Generator();
-    void tick(float time,std::vector<Note> &v,std::vector<Note>&ev);
+    void tick(double time,std::vector<Note> &v,std::vector<Note>&ev);
     bool m_combo_reset = false;
 protected:
 
 private:
     float m_time = 0.0f;
-    void pushNote(float time, int type);
-    void pushEvent(float time, int type);
+    void pushNote(double time, int type);
+    void pushEvent(double time, int type);
     std::vector<float> m_note_times;
     std::vector<int> m_note_types;
     std::vector<float>m_event_times;
